requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
pypdf>=4.0.1
python-dotenv>=1.0.0
streamlit>=1.29.0
google-generativeai>=0.3.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0 